3-show_your_bash_pid_made_easy

+++ 0x05-processes_and_signals/makes_script.sh
@@ -6,7 +6,10 @@ if [ $# -ne 1 ]; then
 	exit
 fi
 
-
+if [ -e "$1" ]; then
+	echo "File exists already"
+	exit
+fi
 echo "#!/usr/bin/env bash" > "$1"
 vi "$1"
 chmod +x "$1"
